Added Widget to warp KeyboardListener and RefreshIndicator
