+++ lib/screens/shockers.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:shock_alarm_app/components/desktop_mobile_refresh_indicator.dart';
 import 'package:shock_alarm_app/screens/shares.dart';
 import 'package:shock_alarm_app/services/alarm_list_manager.dart';
 import 'package:shock_alarm_app/services/openshock.dart';
@@ -314,8 +315,6 @@ class ShockerScreen extends StatefulWidget {
 
 class ShockerScreenState extends State<ShockerScreen> {
   final AlarmListManager manager;
-  final GlobalKey<RefreshIndicatorState> refreshKey = GlobalKey<RefreshIndicatorState>();
-  final FocusNode refreshFocusNode = FocusNode();
 
   void rebuild() {
     setState(() {});
@@ -352,55 +351,39 @@ class ShockerScreenState extends State<ShockerScreen> {
         shockers.add(ShockerItem(shocker: s, manager: manager, onRebuild: rebuild, key: ValueKey(s.getIdentifier())));
       }
     }
-    return KeyboardListener(
-        focusNode: refreshFocusNode,
-        autofocus: true,
-        onKeyEvent: (KeyEvent event) {
-          if (event is KeyDownEvent && event.logicalKey == LogicalKeyboardKey.f5) {
-            refreshKey.currentState?.show();
-          }
-        },
-        child: Column(children: [
-          Text(
-            'All devices',
-            style: t.textTheme.headlineMedium,
-          ),
-          if(groupedShockers.isEmpty)
-            Text(
-                "No shockers found",
-                style: t.textTheme.headlineSmall
-            ),
-          if(!manager.settings.disableHubFiltering)
-            Wrap(spacing: 5,runAlignment: WrapAlignment.start,children: manager.enabledHubs.keys.map<FilterChip>((hub) {
-              return FilterChip(label: Text(manager.getHub(hub)?.name ?? "Unknown hub"), onSelected: (bool value) {
-                manager.enabledHubs[hub] = value;
-                setState(() {
-                }
-                );}, selected: manager.enabledHubs[hub]!);
-            }).toList(),),
-          Flexible(
-              child: RefreshIndicator(
-                  key: refreshKey,
-                  child: ListView(children: [
-                    Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children:
-                        groupedShockers.isNotEmpty ? shockers : []
-                    )
-                  ],),
-                  onRefresh: () async {
-                    await manager.updateShockerStore();
-                    setState(() {});
-                  }
-              )
-          )
-        ],)
-    );
-  }
-
-  @override
-  void dispose() {
-    refreshFocusNode.dispose();
-    super.dispose();
+    return Column(children: [
+      Text(
+        'All devices',
+        style: t.textTheme.headlineMedium,
+      ),
+      if(groupedShockers.isEmpty)
+        Text(
+            "No shockers found",
+            style: t.textTheme.headlineSmall
+        ),
+      if(!manager.settings.disableHubFiltering)
+        Wrap(spacing: 5,runAlignment: WrapAlignment.start,children: manager.enabledHubs.keys.map<FilterChip>((hub) {
+          return FilterChip(label: Text(manager.getHub(hub)?.name ?? "Unknown hub"), onSelected: (bool value) {
+            manager.enabledHubs[hub] = value;
+            setState(() {
+            }
+            );}, selected: manager.enabledHubs[hub]!);
+        }).toList(),),
+      Flexible(
+        child: DesktopMobileRefreshIndicator(
+          onRefresh: () async {
+            await manager.updateShockerStore();
+            setState(() {});
+          },
+          child: ListView(children: [
+            Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children:
+              groupedShockers.isNotEmpty ? shockers : []
+            )
+          ],)
+        )
+      )
+    ],);
   }
 }
